wip: backup: qemu: Wire up qemu full push backup commands over QMP
wip: still need to polish event code, cleanup properly on errors

Update the code to support push backups; for now, the destination
file still has to be local, although the XML could be extended into
supporting remote destinations (where we will have to use the full
power of blockdev-add). This also touches up the event handling
to inform the user when the job is complete.

Signed-off-by: Eric Blake <eblake@redhat.com>