domain: Add VIR_DOMAIN_XML_SNAPSHOTS flag
Right now, copying the state of a transient domain with snapshots
from one host to another requires multiple API calls on both
machines - on the host: get the domain XML, get a list of the
snapshots, and then for each snapshot get the snapshot's XML;
then on the destination: create the domain, then multiple
domain snapshot create calls with the REDEFINE flag.  This
patch aims to make the process use fewer APIs by making it
possible to grab the XML for all snapshots at the same time as
grabbing the domain XML.  Note that we had to do the modification
to virDomainGetXMLDesc(), rather than virDomainSnapshotGetXMLDesc(),
since the latter requires a single non-NULL snapshot object,
whereas we want the list of all snapshots for the domain (even
if the list has 0 elements).

Once wired up in drivers in later patches, the new information
is provided as:

<domain ...>
  ...
  </devices>
  <snapshots current='name'>
    <domainsnapshot>
      ...
    </domainsnapshot>
    <domainsnapshot>
      ...
    </domainsnapshot>
  </snapshots>
</domain>

For now, I did not modify the schema to permit this information
during virDomainDefineXML; it is still necessary to use
virDomainSnapshotCreateXML with VIR_DOMAIN_SNAPSHOT_CREATE_REDEFINE
multiple times to recreate the added state output here.

Unfortunately, libvirt versions between 1.2.12 and 5.0.0 will
silently ignore the new flag, rather than diagnosing that they
don't support it; but at least silent lack of snapshots from
an older server is not a security hole.

Signed-off-by: Eric Blake <eblake@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>