backup: Introduce virDomainCheckpoint APIs
Introduce a bunch of new public APIs related to backup checkpoints.
Checkpoints are modeled heavily after virDomainSnapshotPtr (both
represent a point in time of the guest), although a snapshot exists
with the intent of rolling back to that state, while a checkpoint
exists to make it possible to create an incremental backup at a later
time.

The full list of new API:
        virDomainCheckpointCreateXML;
        virDomainCheckpointCurrent;
        virDomainCheckpointDelete;
        virDomainCheckpointFree;
        virDomainCheckpointGetConnect;
        virDomainCheckpointGetDomain;
        virDomainCheckpointGetParent;
        virDomainCheckpointGetXMLDesc;
        virDomainCheckpointHasMetadata;
        virDomainCheckpointIsCurrent;
        virDomainCheckpointListChildren;
        virDomainCheckpointLookupByName;
        virDomainCheckpointRef;
        virDomainHasCurrentCheckpoint;
        virDomainListCheckpoints;
        virDomainCheckpointGetName;

Signed-off-by: Eric Blake <eblake@redhat.com>